+++ .adonisjs/server/pages.d.ts
@@ -50,5 +50,6 @@ declare module '@adonisjs/inertia/types' {
     'settings/me': ExtractProps<(typeof import('../../inertia/pages/settings/me.vue'))['default']>
     'settings/members': ExtractProps<(typeof import('../../inertia/pages/settings/members.vue'))['default']>
     'settings/org': ExtractProps<(typeof import('../../inertia/pages/settings/org.vue'))['default']>
+    'settings/branding': ExtractProps<(typeof import('../../inertia/pages/settings/branding.vue'))['default']>
   }
 }